changed int to double

+++ src/main/java/com/scalar/producrservice25/Models/Product.java
@@ -9,6 +9,6 @@ public class Product extends BaseModel {
     private String title;
     private String description;
     private String imageUrl;
-    private int price;
+    private double price;
     private Category category;
 }
